+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 # 3.0.0
 * Use custom compiled nginx instead of the packaged version (#20)
 * Rework placeholders syntax in nginx config files
-* Add configurable value for worker_processes setting (default: auto)
+* Add configurable value for worker_processes setting: NGINX_WORK_PROC (default: auto)
 * Change labels org to net.elabftw
 * Group steps together where it makes sense
 * Remove useless VOLUME instructions
+* Add a message on startup with running versions
+* Add a config option to have less messages on startup (SILENT_INIT)
 
 # 2.6.1
 * Modify also the memory_limit value in php.ini